requirements.txt
streamlit
pillow
pyshorteners
